fix log

+++ src/common/common.h
@@ -56,14 +56,14 @@ struct Event {
 using EventList = std::vector<Event>;
 
 struct RAM {
-    int total;
-    int free;
-    int used;
+    unsigned long long total;
+    unsigned long long free;
+    unsigned long long used;
 
     friend void operator>>(const YAML::Node &node, RAM &ram) {
-        ram.total = node["total"].as<int>();
-        ram.free = node["free"].as<int>();
-        ram.used = node["used"].as<int>();
+        ram.total = node["total"].as<unsigned long long>();
+        ram.free = node["free"].as<unsigned long long>();
+        ram.used = node["used"].as<unsigned long long>();
     }
 };
 // GPU struct
